+++ app/Http/Controllers/Api/UserController.php
@@ -33,12 +33,20 @@ class UserController extends Controller
 
     public function update(UserUpdateRequest $request, User $user)
     {
+        if (auth()->id() != $user->id) {
+            return response()->json(['message' => 'operação inválida, login não correspondente'], 401);
+        }
+
         $user->update($request->validated());
         return new UserResource($user);
     }
 
     public function destroy(User $user)
     {
+        if (auth()->id() != $user->id) {
+            return response()->json(['message' => 'operação inválida, login não correspondente'], 401);
+        }
+
         $user->delete();
         return response(null, 204);
     }
